requirements.txt
fastapi
uvicorn
groq
sentence-transformers
faiss-cpu
pandas
python-dotenv
streamlit
requests
